add pagination

+++ FITApp.EmployeesService/Dtos/EmployeesPaginationReedDto.cs
@@ -4,6 +4,6 @@ public class EmployeesPaginationReedDto
 {
     public int Page { get; set; }
     public int PageSize { get; set; }
-    public int TotalCount { get; set; }
+    public long TotalCount { get; set; }
     public IEnumerable<EmployeeDto> Employees { get; set; }
 }
